requirements.txt
google_api_python_client==2.105.0
google_auth_oauthlib==1.1.0
pandas==2.1.1
protobuf==4.24.4
pydantic==2.4.2
streamlit==1.27.2
xmltodict==0.13.0
